+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 * Update :meth:`pyvcf.VcfFrame.chr_prefix`, :meth:`pybed.BedFrame.chr_prefix`, and :meth:`pycov.CovFrame.chr_prefix` methods to skip lines that already have ``chr`` string when ``mode='add'``.
 * Add new methods :meth:`common.rename` and :meth:`pycov.CovFrame.rename`.
 * Add new command :command:`cov-rename`.
+* Add new method :meth:`pyvcf.gt_het`.
 
 0.23.0 (2021-09-21)
 -------------------